Redesign GUI.
This is the first real design of the GUI. This adds
a dashboard, removes the old "items" screen", and adds
new screends for viewing job details and viewing crawl
items. Fixes some issues with job status tracking and
toasts. Adds breadcrumbs to the top nav bar as well as
indicator showing whether the client is connected to
the server.