Allow changing the login URL for a credential
This is an important feature that I overlooked when I first designed
this screen: the login URL is set when you create a new credential,
but I forgot to add a way to change the login URL later on. That is
implemented in this commit. In addition, instead of saving after every
change (e.g. add user, delete user), there is now a single save button
at the bottom, so the user can make several changes and then save them
all at once.